% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\name{add_quotations}
\alias{add_quotations}
\title{duplicated functionality to d8ahelper::copy_unique(quotes = TRUE)}
\usage{
add_quotations(string)
}
\description{
duplicated functionality to d8ahelper::copy_unique(quotes = TRUE)
}
