% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{add_wmy}
\alias{add_wmy}
\title{Add week, weekday, month, year columns based on datetime column}
\usage{
add_wmy(df, dt_col)
}
\arguments{
\item{dt_col}{}
}
\description{
Add week, weekday, month, year columns based on datetime column
}
