% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{any_dups}
\alias{any_dups}
\title{Output summary of duplicated rows by given keys}
\usage{
any_dups(df, keys)
}
\arguments{
\item{keys}{a character vector contains column names to be used for grouping}
}
\value{
a list, of which 'dups' is a subset dataframe contain only duplicated entries by group, 'dup_count' is a count summary by group, 'dup_row' is row index of duplicated rows by group, and 'dup_row_bool' is boolean of duplicated rows by group
}
\description{
Output summary of duplicated rows by given keys
}
