% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{clean_by_id}
\alias{clean_by_id}
\title{Remove rows where id columns are all missing, and rows where all columns but the id columns are missing; (optional) Remove columns where all rows are of missing values.}
\usage{
clean_by_id(df, id_col, var_col = NULL, filter_col = FALSE, th = 1)
}
\arguments{
\item{id_col}{numeric vector specify which columns are considered id columns}

\item{var_col}{numeric vector specify which columns are considered data columns}

\item{filter_col}{a bool, if TRUE then will leave only columns that does not contain missing - applies to all non id columns if var_col is NULL otherwise only applies to var_col only.}

\item{th}{a numeric, range from 0-1 to specify the threshold for amount least amount of non-missing data by column in fractions of total rows}
}
\description{
Remove rows where id columns are all missing, and rows where all columns but the id columns are missing; (optional) Remove columns where all rows are of missing values.
}
