% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{clear_dir}
\alias{clear_dir}
\title{given a directory, delete files given the regex pattern excluding files matching to 'avoid' arg}
\usage{
clear_dir(dir, pattern = ".csv", recursive = F, avoid = "^_")
}
\arguments{
\item{pattern}{a regex string, used to match files to be deleted}

\item{avoid}{a regular expression string, matched files will not be deleted}
}
\description{
given a directory, delete files given the regex pattern excluding files matching to 'avoid' arg
}
