% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalescence join x, y data frames, for columns of same names, append y values to x if rows that are missing value}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  join = dplyr::left_join,
  ...
)
}
\description{
Coalescence join x, y data frames, for columns of same names, append y values to x if rows that are missing value
}
