% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{contain_value}
\alias{contain_value}
\title{examine if a data holder contains 'value' in a pre-defined term}
\usage{
contain_value(x)
}
\arguments{
\item{x}{a vector or a list}
}
\description{
examine if a data holder contains 'value' in a pre-defined term
}
