% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_time_to_chr}
\alias{convert_time_to_chr}
\title{Convert time related columns to character}
\usage{
convert_time_to_chr(df, regex = "Time")
}
\arguments{
\item{df}{a data frame}

\item{regex}{a regular expression character vector, to be used for matching to 'time' columns}
}
\description{
background: POSIXct datetime format when output as csv will not be read corretly with JMP \cr
purpose: convert all time columns to character format befor exporting to csv \cr
note: accept regex if default rule of selecting columns containing "time" is not ideal \cr
}
