% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{copy_as_sql}
\alias{copy_as_sql}
\title{SQL friendly formating tp combine list of inputs}
\usage{
copy_as_sql(x = paste(readClipboard(), collapse = ","),
  load_csv = FALSE, col = "step")
}
\description{
SQL friendly formating tp combine list of inputs
}
