% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{copy_as_sql_like}
\alias{copy_as_sql_like}
\title{SQL friendly formating tp combine list of inputs with 'LIKE' clause}
\usage{
copy_as_sql_like(x = paste(readClipboard(), collapse = ","),
  is_lotid = TRUE, col = "lot_id", add_and_prior = FALSE,
  add_and_post = FALSE)
}
\description{
SQL friendly formating tp combine list of inputs with 'LIKE' clause
}
