% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{corr_to_df}
\alias{corr_to_df}
\title{ilter high corr features and output in tidy format}
\usage{
corr_to_df(cor, level = 0.9)
}
\arguments{
\item{cor}{corr matrix of numerical format}

\item{level}{}
}
\description{
ilter high corr features and output in tidy format
}
