% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{count_lines}
\alias{count_lines}
\title{Count number of lines without using read_csv}
\usage{
count_lines(filepath, batch = 1000)
}
\arguments{
\item{batch}{}
}
\description{
Count number of lines without using read_csv
}
