% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{create_dummy}
\alias{create_dummy}
\title{wraper function to use caret::dummyVars to create dummy variables from character variables}
\usage{
create_dummy(df, response, simplify = FALSE, filter_chr = TRUE, ...)
}
\arguments{
\item{df}{a data frame}

\item{response}{response columns, if any}

\item{simplify}{a logical for if to simplify output to a dataframe, if FALSE returns a list}

\item{filter_chr}{a logical for if calling filter_col_unique for removing columns based on near zero variance predictors}
}
\description{
wraper function to use caret::dummyVars to create dummy variables from character variables
}
