% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\name{create_waferid}
\alias{create_waferid}
\title{if 'lotid' or 'startlot' column exist along with alias column, create 'waferid' column}
\usage{
create_waferid(df, waferid2 = TRUE)
}
\description{
note:
(by default), create 'waferid2'column (unique wafer id in the format of 'startlot-alias')
}
