% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{days_to_date}
\alias{days_to_date}
\title{Use number of days past 1970-01-01 to indicate date}
\usage{
days_to_date(days)
}
\arguments{
\item{days}{}
}
\description{
Use number of days past 1970-01-01 to indicate date
}
