% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{decode_col}
\alias{decode_col}
\title{Sister func to encode_col
returns a data frame with original column names based on df$dict}
\usage{
decode_col(df.list, df = df.list$df)
}
\arguments{
\item{df.list:}{list object generated from d8ahelper::convert_col_name}

\item{df:}{data frame to convert names from, default to df within df.list}
}
\description{
Sister func to encode_col
returns a data frame with original column names based on df$dict
}
