% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{decrypt}
\alias{decrypt}
\title{A wrapper function for decrypting/delete file using key}
\usage{
decrypt(file, dest, key = gen_key(), keep.original = FALSE)
}
\description{
A wrapper function for decrypting/delete file using key
}
