% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{encode_col}
\alias{encode_col}
\title{Replace column names with alpha-numeric sequences, returns a named vector (of name-value pairs) for column name look-up; Returns a list.}
\usage{
encode_col(df)
}
\description{
Replace column names with alpha-numeric sequences, returns a named vector (of name-value pairs) for column name look-up; Returns a list.
}
