% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{encrypt}
\alias{encrypt}
\title{A wrapper function to encrypt file using a key}
\usage{
encrypt(file_orig, file_enc, key_file)
}
\arguments{
\item{key_file}{}
}
\description{
A wrapper function to encrypt file using a key
}
