% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\name{expand_wafer_list}
\alias{expand_wafer_list}
\title{expand lot level table to a wafer level table}
\usage{
expand_wafer_list(lotlist = paste(readClipboard(), collapse = ","))
}
\description{
input: (by default) jmp table lot id column being coppied or other forms of lotid string
}
\details{
output:
a dataframe with lotid column and all waferids from 01-25
copy unique wafer ids as comma seperated string to clipboard
}
