% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{file.copy.content.only}
\alias{file.copy.content.only}
\title{R's base::file.copy can only copy a directory as a entirety to another directory
to enable copy files to another directory
allow re matching to copy only a subset of all files}
\usage{
file.copy.content.only(
  from,
  to,
  pattern = ".*",
  overwrite = FALSE,
  copy.mode = TRUE,
  copy.date = FALSE
)
}
\description{
R's base::file.copy can only copy a directory as a entirety to another directory
to enable copy files to another directory
allow re matching to copy only a subset of all files
}
