% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fill_as_na}
\alias{fill_as_na}
\title{Sister func to fill_na_as_missing, replace cells with NAs if match to given string}
\usage{
fill_as_na(df, pattern = "[missing]")
}
\description{
Sister func to fill_na_as_missing, replace cells with NAs if match to given string
}
