% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fill_cols}
\alias{fill_cols}
\title{find columns that has at minimum 'th' number of non-NA values and fill missings with upper above closest non-NA value}
\usage{
fill_cols(df, th = 1, direction = "down")
}
\arguments{
\item{th}{threshold value used to identify columns with minimal number of unique value count, if min value < th then add to target column list}

\item{direction}{}
}
\description{
find columns that has at minimum 'th' number of non-NA values and fill missings with upper above closest non-NA value
}
