% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fill_map}
\alias{fill_map}
\title{a wrapper function to 'gstat::gstat' function to use a univariate or multivariate geostatistical model with input of x,y coordinate and fill z column with predictions, for each id group.
puff_my_df can be used to make all available entries available for each group}
\usage{
fill_map(df, idcol, xcol = "x", ycol = "y", zcol = "z")
}
\description{
a wrapper function to 'gstat::gstat' function to use a univariate or multivariate geostatistical model with input of x,y coordinate and fill z column with predictions, for each id group.
puff_my_df can be used to make all available entries available for each group
}
