% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fill_missing_as_na}
\alias{fill_missing_as_na}
\title{Fill literal 'missing' or 'NA' with NAs}
\usage{
fill_missing_as_na(
  df,
  fill = NA,
  pat = "^(\\\\s*\\\\[missing\\\\]\\\\s*|\\\\s*|NA|)$"
)
}
\arguments{
\item{pat}{}
}
\description{
Fill literal 'missing' or 'NA' with NAs
}
