% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{filter_col_unique}
\alias{filter_col_unique}
\title{wraper function to caret::nearZeroVar to filter chr columns for Zero- and Near Zero-Variance Predictors}
\usage{
filter_col_unique(df, percentUnique = 50, freqRatio = 1, breaks = 50,
  useDefault = FALSE, plot = TRUE, simplify = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{percentUnique}{a numeric for percentUnique threshold}

\item{freqRatio}{a numeric for freqRatio threshold}

\item{useDefault}{a logical for if to use caret default for removing near zero variance variables}

\item{plot}{a logical for if to include histograms of percentUnique and freqRatio}

\item{simplify}{a logical for if to simplify output to a single data frame, else return a list}
}
\description{
wraper function to caret::nearZeroVar to filter chr columns for Zero- and Near Zero-Variance Predictors
}
