% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\name{format_col}
\alias{format_col}
\title{remove from this package, for now
format column names to make them more readable.
needs a revisit}
\usage{
format_col(df)
}
\description{
remove from this package, for now
format column names to make them more readable.
needs a revisit
}
