% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_datetime}
\alias{format_datetime}
\title{Format datetime columns to a specified format}
\usage{
format_datetime(df, regex, timeformat = "\%Y/\%m/\%d \%H:\%M:\%S \%p", ...)
}
\arguments{
\item{df}{a data frame}

\item{regex}{a character used for regular expression lookup of matching columns to format with}

\item{timeformat}{datetime format to be used for formating columns with (see 'help(strptime)')}
}
\description{
Format datetime columns to a specified format
}
