% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\name{format_lotid}
\alias{format_lotid}
\title{format lot ids
accept less than 7 characters format (usually numerical)
if passed in startlot ids, will return lotids with '.002'}
\usage{
format_lotid(lot)
}
\description{
format lot ids
accept less than 7 characters format (usually numerical)
if passed in startlot ids, will return lotids with '.002'
}
