% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_to_percentage}
\alias{format_to_percentage}
\title{Helper functions for general data processing at IM Flash.
given a integer/floating point number, convert to 'dd.dd%' percentageformat}
\usage{
format_to_percentage(x, digits = 2, format = "f", ...)
}
\description{
Helper functions for general data processing at IM Flash.
given a integer/floating point number, convert to 'dd.dd%' percentageformat
}
