% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{fread2}
\alias{fread2}
\title{wrapper function to data.table::fread to convert blank cells to NA at reading
for some reason data.table::fread do not convert blank cell to NA even with na.string = ""
see https://stackoverflow.com/questions/51019041/blank-space-not-recognised-as-na-in-fread}
\usage{
fread2(file, ...)
}
\description{
wrapper function to data.table::fread to convert blank cells to NA at reading
for some reason data.table::fread do not convert blank cell to NA even with na.string = ""
see https://stackoverflow.com/questions/51019041/blank-space-not-recognised-as-na-in-fread
}
