% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{gen_key}
\alias{gen_key}
\title{A wrapper function for generating, store, and read key}
\usage{
gen_key(file = ".key", key.wrap = TRUE)
}
\description{
A wrapper function for generating, store, and read key
}
