% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{headtail}
\alias{headtail}
\title{quick glance into a dataframe combining n head and tail rows}
\usage{
headtail(df, n = 5)
}
\arguments{
\item{n}{number of top/bottom rows to include}
}
\value{
top, bottom and a transition rows suggest how many rows skipped
}
\description{
quick glance into a dataframe combining n head and tail rows
}
\examples{
headtail(mtcars)
}
