% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{insp_dir}
\alias{insp_dir}
\title{Perform a non intrusive inspection of a given directory and return a dataframe with essential file information sorted by file size}
\usage{
insp_dir(dir, recursive = TRUE, pattern = ".*", silence = TRUE)
}
\arguments{
\item{silence}{}
}
\description{
Perform a non intrusive inspection of a given directory and return a dataframe with essential file information sorted by file size
}
