% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{is_fct_or_chr}
\alias{is_fct_or_chr}
\title{check if is factor or character}
\usage{
is_fct_or_chr(x)
}
\arguments{
\item{x}{}
}
\description{
check if is factor or character
}
\examples{
is_fct_or_chr(as.factor(c(1,2,3,3,4,2,2,1)))
is_fct_or_chr(c("A","A","B","c"))
}
