% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{lapply_preserve_names}
\alias{lapply_preserve_names}
\title{lapply wrapper but preserve object names as the name suggests}
\usage{
lapply_preserve_names(list, fun)
}
\arguments{
\item{fun}{}
}
\description{
lapply wrapper but preserve object names as the name suggests
}
