% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{list.files2}
\alias{list.files2}
\title{wraper function list.files to allow coppied windows path format containing '/'}
\usage{
list.files2(path = readr::clipboard(), return_str = TRUE, ...)
}
\description{
wraper function list.files to allow coppied windows path format containing '/'
}
