% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{load_csv}
\alias{load_csv}
\title{Fast way to load CSV files}
\usage{
load_csv(file.name = "R_output.csv", path = here::here("R_output"),
  folder = "", load = TRUE)
}
\arguments{
\item{file.name}{character name for output file name}

\item{path}{target folder, by default use 'R_output' folder under root directory}

\item{folder}{specify output folder under 'path'}

\item{load}{a boolean value, if == "FALSE" then only print to console the existing files}
}
\description{
Fast way to load CSV files
}
