% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{load_files}
\alias{load_files}
\title{load files paths in specified location}
\usage{
load_files(
  load_from,
  pattern = "\\\\.csv$",
  read_as_df = TRUE,
  use_fread = TRUE,
  avoid = "^_|archive",
  full.names = TRUE,
  load = FALSE,
  ...
)
}
\arguments{
\item{load_from}{a path for where files are load from}

\item{read_as_df}{a logical, if to read as data frame, else read as text}

\item{load}{a logical for if actually read the files; if pattern is not ".csv", then read as text}
}
\description{
load files paths in specified location
}
