% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{map_preserve_names_use_future}
\alias{map_preserve_names_use_future}
\title{map but preserve object names, use future to add parallel processing}
\usage{
map_preserve_names_use_future(list, fun)
}
\arguments{
\item{fun}{}
}
\description{
map but preserve object names, use future to add parallel processing
}
