% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{missing_as_na}
\alias{missing_as_na}
\title{Sister func to na_as_missing, replace cells with NAs if match to given string}
\usage{
missing_as_na(df, pattern = "-")
}
\description{
Sister func to na_as_missing, replace cells with NAs if match to given string
}
