% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{model_lm}
\alias{model_lm}
\title{Run lm model and print a curated summary output}
\usage{
model_lm(formula, dataset)
}
\arguments{
\item{dataset}{}
}
\description{
Run lm model and print a curated summary output
}
