% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{move_left}
\alias{move_left}
\title{Move column or columns to far left hand side of a dataframe}
\usage{
move_left(df, str)
}
\arguments{
\item{df}{a data frame}

\item{str}{column name(s)}
}
\description{
Accept multiple columns if passed in a character vector containing column names
}
