% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{multi_join}
\alias{multi_join}
\title{sister function to coalesce_join, join together a list of data frames}
\usage{
multi_join(list, by = NULL, join = dplyr::full_join, ...)
}
\description{
sister function to coalesce_join, join together a list of data frames
}
