% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{multiplot}
\alias{multiplot}
\title{list of ggplot objects into one plot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
list of ggplot objects into one plot
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom
}
