% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{na_as_missing}
\alias{na_as_missing}
\title{Identify NAs and empty cells that fit a pattern and replace}
\usage{
na_as_missing(
  df,
  pat = "^(\\\\s*\\\\[*missing\\\\]*\\\\s*|\\\\s*|NA|\\\\s*-\\\\s*)$",
  fill = "-"
)
}
\arguments{
\item{fill}{}
}
\description{
Identify NAs and empty cells that fit a pattern and replace
}
