% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{open_encrypted}
\alias{open_encrypted}
\title{A wrapper function for the workflow of decryption, source and encryption of a '.R' file}
\usage{
open_encrypted(file_actual, file_encrypted, encrypt = FALSE)
}
\description{
A wrapper function for the workflow of decryption, source and encryption of a '.R' file
}
