% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{boxplot with confidence intervals and summary statistics such as p-val and risk}
\usage{
plot_boxplot(
  num,
  grp = NULL,
  data = NULL,
  anova = TRUE,
  alpha = 0.05,
  verbose = TRUE,
  plot_only = TRUE,
  ...
)
}
\description{
boxplot with confidence intervals and summary statistics such as p-val and risk
}
