% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{remove_dup_cols}
\alias{remove_dup_cols}
\title{Taken a list of data frames, remove duplicazte columns that exist across multiple dataframes}
\usage{
remove_dup_cols(df_list)
}
\arguments{
\item{df_list}{}
}
\description{
Taken a list of data frames, remove duplicazte columns that exist across multiple dataframes
}
