% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{remove_high_corr_features}
\alias{remove_high_corr_features}
\title{wraper function ofcaret::findCorrelation: find and remove highly correlated features
returns a list of 1) reduced df 2) corr table of removed features
requires data frame of numeric w/o missing values}
\usage{
remove_high_corr_features(df, level = 0.99)
}
\description{
wraper function ofcaret::findCorrelation: find and remove highly correlated features
returns a list of 1) reduced df 2) corr table of removed features
requires data frame of numeric w/o missing values
}
