% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rm_dup_cols_in_a_list}
\alias{rm_dup_cols_in_a_list}
\title{Taken a list of data frames, check each column and remove duplicated ones from each dataframe if they have already been identified in LHS data frames in the list}
\usage{
rm_dup_cols_in_a_list(df_list, exclude = c())
}
\arguments{
\item{df_list}{a list of data.frame objects}

\item{exclude}{names of columns that will be left alone from being removed}
}
\description{
Taken a list of data frames, check each column and remove duplicated ones from each dataframe if they have already been identified in LHS data frames in the list
}
