% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rm_dups_w_less_data}
\alias{rm_dups_w_less_data}
\title{Remove duplicated rows by leaving only 1 observation per each group of most non-missing data columns
if multiple observations sharing equal number of most non-missing data columns, choose the first one by row index
a wrapper function of d8ahelper::any_dups}
\usage{
rm_dups_w_less_data(df, keys)
}
\arguments{
\item{keys}{a character vector contains column names, for grouping purposes}
}
\description{
Remove duplicated rows by leaving only 1 observation per each group of most non-missing data columns
if multiple observations sharing equal number of most non-missing data columns, choose the first one by row index
a wrapper function of d8ahelper::any_dups
}
