% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rm_single_unique_col}
\alias{rm_single_unique_col}
\title{Remove column(s) that contain a single unique value}
\usage{
rm_single_unique_col(df, threshold = 1, na_as_a_cat = TRUE, exclude = c())
}
\description{
Remove column(s) that contain a single unique value
}
