% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{save_csv}
\alias{save_csv}
\title{Fast way to output CSV files}
\usage{
save_csv(df, time_as_chr = FALSE, path = here::here("R_output"),
  file.name = "R_output.csv", folder = "")
}
\arguments{
\item{df}{a data frame}

\item{time_as_chr}{specify if to convert datetime columns to character before export}

\item{path}{ouput folder, by default use 'R_output' folder under root directory}

\item{file.name}{character name for output file name}

\item{folder}{specify output folder under 'path'}
}
\description{
Fast way to output CSV files
}
