% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{secs_to_date}
\alias{secs_to_date}
\title{Convert numerically converted format POSIXct time back to POSIXct datetime format}
\usage{
secs_to_date(secs)
}
\description{
Convert numerically converted format POSIXct time back to POSIXct datetime format
}
