% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{str_find}
\alias{str_find}
\title{wrapper function for locate a string based on regex leading to and after it}
\usage{
str_find(str, start, end, replacement, return_loc = FALSE)
}
\arguments{
\item{str}{a string to search from}

\item{start}{a regex string that lead to the string of interest}

\item{end}{a regex string that is after the string of interest}

\item{replacement}{a string to replace the located string, optional.}
}
\value{
_loc a logical, if TRUE then return character start/end location instead of the string itself
}
\description{
wrapper function for locate a string based on regex leading to and after it
}
