% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{subset_by_quantile}
\alias{subset_by_quantile}
\title{Filter column by removing points beyond top and(or) bottom percentage thresholds}
\usage{
subset_by_quantile(df, col, top = 0, bottom = 0, ...)
}
\arguments{
\item{...}{}
}
\description{
Filter column by removing points beyond top and(or) bottom percentage thresholds
}
