% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_col}
\alias{sum_col}
\title{Summarize a data frame with concise and useful summary statistics}
\usage{
sum_col(df, outlier_method = "z", thres = 3, ...)
}
\arguments{
\item{df}{a data frame}

\item{outlier_method}{"z" (Normal) or "mad" (Nonparametric) for how to categorize outliers}

\item{thres}{a number to suggest number of standard deviations to use for categorize outliers}
}
\value{
a data frame
}
\description{
Summarize a data frame with concise and useful summary statistics
}
