% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_df}
\alias{sum_df}
\title{Sister function to sum_col and sum_row, returns a list of results from each}
\usage{
sum_df(df)
}
\description{
Sister function to sum_col and sum_row, returns a list of results from each
}
