% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_missing}
\alias{sum_missing}
\title{give top n missing value column combinations}
\usage{
sum_missing(df, num = 5, ...)
}
\arguments{
\item{df}{a data frame}

\item{num}{top n number of missing rows to return}
}
\description{
give top n missing value column combinations
}
