% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_table}
\alias{sum_table}
\title{Given 2 columns, summarize counts of each unique value-pair combinations}
\usage{
sum_table(df, x, y, sum = TRUE, ...)
}
\arguments{
\item{x}{a column #, unique values displayed as row headers}

\item{y}{a column #, unique values displayed as column headers}
}
\value{
a data frame
}
\description{
Given 2 columns, summarize counts of each unique value-pair combinations
}
