% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{trim_ws}
\alias{trim_ws}
\title{Remove leading and trailing white spaces}
\usage{
trim_ws(x, leading = FALSE, trailing = FALSE, all = FALSE, replacement = "_")
}
\arguments{
\item{x}{a character}

\item{all}{a boolean, trim non-leading/trailing white spaces as well if equals TRUE}

\item{replacement}{a character to replace non-leading/trailing white spaces with}
}
\description{
By default, trim both leading and traling white spaces unless specified
}
