% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{trim_ws_df}
\alias{trim_ws_df}
\title{Remove leading and trailing whitespaces that are within any character or factor type columns of a data.frame}
\usage{
trim_ws_df(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Remove leading and trailing whitespaces that are within any character or factor type columns of a data.frame
}
