% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\name{unique_lotids}
\alias{unique_lotids}
\title{yield3 like lot id formating}
\usage{
unique_lotids(lotlist = paste(readClipboard(), collapse = ","))
}
\description{
input: (by default) jmp table lot id column being coppied or other forms of lotid string
}
\details{
output:
a data.frame of startlot/lotid columns
and copy comma seperated unique lot ids to clipboard
}
