% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\name{whole_shabang}
\alias{whole_shabang}
\title{the whole shabang of formatting lotid/waferid columns
(by default) remove columns with single unique value}
\usage{
whole_shabang(df, col_remove = TRUE)
}
\description{
the whole shabang of formatting lotid/waferid columns
(by default) remove columns with single unique value
}
