% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{add_empty_rows}
\alias{add_empty_rows}
\title{Append empty rows to a dataframe to make row number = n}
\usage{
add_empty_rows(df, n)
}
\description{
Append empty rows to a dataframe to make row number = n
}
