% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalescence join lhs x and rhs y data frames, for each column-row combination, append rhs to lhs if missing, use lhs value if there are conflicts}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  join = dplyr::left_join,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Coalescence join lhs x and rhs y data frames, for each column-row combination, append rhs to lhs if missing, use lhs value if there are conflicts
}
