% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Join x, y data frames by appending y values to x if missing (for columns of same names)}
\usage{
coalesce_join(x, y, by = NULL, suffix = c(".x", ".y"),
  join = dplyr::left_join, ...)
}
\description{
Join x, y data frames by appending y values to x if missing (for columns of same names)
}
