% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{copy_unique}
\alias{copy_unique}
\title{Copy unique values to clipboard.}
\usage{
copy_unique(
  df = paste(readClipboard(), collapse = ","),
  col = "value",
  sep = ",",
  reduce_only = FALSE,
  fix_len = FALSE,
  quotes = FALSE,
  save_csv = FALSE,
  ...
)
}
\arguments{
\item{df}{a character string copied from clipboard, or a dataframe column}

\item{col}{column name if df is a data frame}
}
\value{
a comma separated string w/ duplicates removed and a dataframe of unique values
}
\description{
\enumerate{
\item if input is a dataframe, do not return back a df \cr
\item by default will assume lotids being copied and format them unless set format_lotid = FALSE \cr
\item if format_step == TRUE, return a comma separated step name string instead \cr
\item if save_csv == TRUE, will save output to '1.lot_list.csv' \cr
}
}
