% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{corr_to_df}
\alias{corr_to_df}
\title{filter high corr features and output in tidy format
input: corr matrix of numerical format
output: a data frame}
\usage{
corr_to_df(cor, level = 0.9)
}
\description{
filter high corr features and output in tidy format
input: corr matrix of numerical format
output: a data frame
}
