% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{decrypt}
\alias{decrypt}
\title{A wrapper function to decrypt file using a key}
\usage{
decrypt(file_enc, file_orig, key_file)
}
\arguments{
\item{key_file}{}
}
\description{
A wrapper function to decrypt file using a key
}
