% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{encode_col}
\alias{encode_col}
\title{replace column names with alpha-numeric sequences
returns a named vector (of name-value pairs) for column name look-up
returns a list}
\usage{
encode_col(df)
}
\description{
replace column names with alpha-numeric sequences
returns a named vector (of name-value pairs) for column name look-up
returns a list
}
