% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{encrypt}
\alias{encrypt}
\title{A wrapper function for encrypting/delete file using key}
\usage{
encrypt(file, dest, key = gen_key(), keep.original = FALSE)
}
\description{
A wrapper function for encrypting/delete file using key
}
