% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{file.copy.content.only}
\alias{file.copy.content.only}
\title{copy files to another directory}
\usage{
file.copy.content.only(
  from,
  to,
  pattern = ".*",
  overwrite = FALSE,
  copy.mode = TRUE,
  copy.date = FALSE
)
}
\arguments{
\item{copy.date}{}
}
\description{
copy files to another directory
}
