% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_num}
\alias{format_num}
\title{Format numbers to limit digits after decimal point}
\usage{
format_num(num, digits = 2, ...)
}
\arguments{
\item{...}{}
}
\description{
Accept numerical vectors, or data.frame
}
