% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_to_percentage}
\alias{format_to_percentage}
\title{Convert integer/floating point number format to 'dd.dd%' percentage format}
\usage{
format_to_percentage(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{a number}

\item{digits}{the desired number of digits after the decimal point}

\item{format}{default to "f" which gives numbers in the usual xxx.xxx format. see help(formatC)}
}
\description{
Convert integer/floating point number format to 'dd.dd%' percentage format
}
