% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{fread2}
\alias{fread2}
\title{wrapper function to data.table::fread to convert blank cells to NA at reading
https://stackoverflow.com/questions/51019041/blank-space-not-recognised-as-na-in-fread}
\usage{
fread2(file, ...)
}
\arguments{
\item{...}{}
}
\description{
wrapper function to data.table::fread to convert blank cells to NA at reading
https://stackoverflow.com/questions/51019041/blank-space-not-recognised-as-na-in-fread
}
