% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{gen_key}
\alias{gen_key}
\title{A wrapper function for sodium::keygen to generate, convert keys}
\usage{
gen_key(key_file, convert_to_sodium_key = FALSE)
}
\arguments{
\item{convert_to_sodium_key}{}
}
\description{
A wrapper function for sodium::keygen to generate, convert keys
}
