% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_id}
\alias{get_id}
\title{get id from encoded obj, sister func to encode_col, retrieve column id based on names}
\usage{
get_id(df, ...)
}
\arguments{
\item{...}{}
}
\description{
get id from encoded obj, sister func to encode_col, retrieve column id based on names
}
\examples{
get_id(encode_col(mtcars),'mpg')
get_id(encode_col(mtcars),c('mpg','disp'))
}
