% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_id}
\alias{get_id}
\title{sister function to encode_col, retrieve column id based on name(s)}
\usage{
get_id(df, ...)
}
\description{
sister function to encode_col, retrieve column id based on name(s)
}
