% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_name}
\alias{get_name}
\title{sister function to encode_col, retrieve column names based on id(s)}
\usage{
get_name(df, ...)
}
\description{
sister function to encode_col, retrieve column names based on id(s)
}
