% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_name}
\alias{get_name}
\title{get names of encoded obj, sister func to encode_col, retrieve column names based on ids}
\usage{
get_name(df, ...)
}
\arguments{
\item{...}{}
}
\description{
get names of encoded obj, sister func to encode_col, retrieve column names based on ids
}
