% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{headtail}
\alias{headtail}
\title{A quick glance into a dataframe combining n head and tail rows}
\usage{
headtail(df, n = 5)
}
\arguments{
\item{n}{number of top/bottom rows to include, or number of characters in case of a string}
}
\value{
top, bottom and a transition rows suggest how many rows skipped
}
\description{
A quick glance into a dataframe combining n head and tail rows
}
\examples{
headtail(mtcars)
headtail("I would like to make this a very long string in order to make my case")
}
