% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{insert_nas}
\alias{insert_nas}
\title{Insert NAs randomly to a data.frame}
\usage{
insert_nas(df, percent = 0.2)
}
\arguments{
\item{df}{a data frame}

\item{percent}{percent of NAs per column}
}
\value{
a data frame
}
\description{
Insert NAs randomly to a data.frame
}
