% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{lapply_preserve_names}
\alias{lapply_preserve_names}
\title{lapply wrapper, but preserve object names}
\usage{
lapply_preserve_names(list, fun)
}
\description{
base lapply function give access to only the element of the vector but not other attributes
such as name
}
