% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{lapply_preserve_names}
\alias{lapply_preserve_names}
\title{lapply wrapper but preserve object names as the name suggests}
\usage{
lapply_preserve_names(list, fun)
}
\arguments{
\item{fun}{}
}
\description{
lapply wrapper but preserve object names as the name suggests
}
\examples{
lapply_preserve_names(list(first.item=c(1,2,3), second.item=c(3,2,1)), function(x)paste0("each list member's name is accessible now! --> ", names(x)))
}
