% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{list_files_fwd_slash}
\alias{list_files_fwd_slash}
\title{wrapper function list.files to allow coppied windows path format containing backward slashes}
\usage{
list_files_fwd_slash(path = readr::clipboard(), return_str = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
wrapper function list.files to allow coppied windows path format containing backward slashes
}
