% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{load_files}
\alias{load_files}
\title{load files paths in specified location}
\usage{
load_files(loc, pattern = ".csv", avoid = "^_|archive",
  full.names = TRUE, load = FALSE, ...)
}
\description{
load files paths in specified location
}
