% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{move_left}
\alias{move_left}
\title{Move column to far left hand side}
\usage{
move_left(df, str)
}
\arguments{
\item{df}{a data frame}

\item{str}{column name(s)}
}
\description{
Accept multiple columns if passed in a character vector containing column names
}
\examples{
View(move_left(mtcars, "wt"))
View(move_left(mtcars, c("gear", "carb", "wt")))
}
