% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rm_single_unique_col}
\alias{rm_single_unique_col}
\title{Remove column(s) if containing a single unique value}
\usage{
rm_single_unique_col(df, threshold = 1)
}
\description{
Remove column(s) if containing a single unique value
}
