% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{save_csv}
\alias{save_csv}
\title{A custom way to output CSV files}
\usage{
save_csv(
  df,
  file.name = "r_output.csv",
  path = here::here("r_output"),
  folder = "",
  overwrite = FALSE,
  use_readr = FALSE,
  as_chr = TRUE,
  ...
)
}
\arguments{
\item{df}{a data frame}

\item{file.name}{character name for output file name}

\item{path}{ouput folder, by default use 'r_output' folder under root directory}

\item{folder}{specify output folder under 'path'}

\item{time_as_chr}{specify if to convert datetime columns to character before export}
}
\description{
Will copy file path to clipboard for convinient file locate or remove (via unlink())
}
