% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{save_csv_from_a_list}
\alias{save_csv_from_a_list}
\title{A wrapper function of save_csv that applies to a list of data frames}
\usage{
save_csv_from_a_list(
  lst,
  overwrite = FALSE,
  use_name = FALSE,
  use_clean_names = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
A wrapper function of save_csv that applies to a list of data frames
}
