% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_missing}
\alias{sum_missing}
\title{Summary statistics of top n missing value columns}
\usage{
sum_missing(df, num = 5, ...)
}
\arguments{
\item{df}{a data frame}

\item{num}{top n number of missing rows to return}
}
\description{
Summary statistics of top n missing value columns
}
